+++ CHANGELOG.md
@@ -604,3 +604,10 @@
 * Define `U_III_UU_I_II` data type
 * Define `W_II_II_I` data type
 * Define `W_I_II_I` data type
+* Define `Automation` type alias
+* Define `Transition'` type alias
+* Define `State` type alias
+* Define `Statefully'` pattern
+* Define `Transition'` pattern
+* Define `switch` expression
+* Define `review` expression